db_handler/requirements.txt
kafka-python==1.4.7
psycopg2-binary==2.9.3
sqlalchemy==1.4.31